Feat/perf and mem (#50)
This improves the perf of the lib generally by removing the Argon2Memory class and using System.Memory and spans in its place. This prevents allocations in the hot loop and makes things work a little bit nicer. Also blitting was simplified and made into an extension on Memory which works out a bit nicer.

I've added a benchmark suite and a comparison harness to ensure that things are still working:

The results of the benchmark are below (Note the before and after). As iterations increase the memory savings becomes larger and larger.

I also fixed a bug where the blake initializer was never called and always assigned to the default. I believe that's #42
You'll notice that there are a few formatting changes that were just victims of my "format current file" muscle memory.

Let me know if there's any changes you'd like made or any concerns and I'll try to get to them pretty soon. I'll get to some cleanup this weekend as well

closes #48
closes #42